Move 'AddContentDirective' to a dedicated module
And update mypy usage to ensure that type checking passes on the newly
isolated class.

In particular, update typing dependencies to include `sphinx`.
