+++ requirements/py3.13/typing.txt
@@ -4,25 +4,59 @@
 #
 #    tox p -m freezedeps
 #
-certifi==2024.8.30
+alabaster==1.0.0
+    # via sphinx
+babel==2.16.0
+    # via sphinx
+certifi==2025.1.31
     # via requests
-charset-normalizer==3.4.0
+charset-normalizer==3.4.1
     # via requests
+docutils==0.21.2
+    # via sphinx
 idna==3.10
     # via requests
-mypy==1.12.0
+imagesize==1.4.1
+    # via sphinx
+jinja2==3.1.5
+    # via sphinx
+markupsafe==3.0.2
+    # via jinja2
+mypy==1.14.1
     # via -r .typing.in
 mypy-extensions==1.0.0
     # via mypy
+packaging==24.2
+    # via sphinx
+pygments==2.19.1
+    # via sphinx
 pyyaml==6.0.2
     # via responses
 requests==2.32.3
-    # via responses
-responses==0.25.3
+    # via
+    #   responses
+    #   sphinx
+responses==0.25.6
     # via -r .typing.in
+snowballstemmer==2.2.0
+    # via sphinx
+sphinx==8.1.3
+    # via -r .typing.in
+sphinxcontrib-applehelp==2.0.0
+    # via sphinx
+sphinxcontrib-devhelp==2.0.0
+    # via sphinx
+sphinxcontrib-htmlhelp==2.1.0
+    # via sphinx
+sphinxcontrib-jsmath==1.0.1
+    # via sphinx
+sphinxcontrib-qthelp==2.0.0
+    # via sphinx
+sphinxcontrib-serializinghtml==2.0.0
+    # via sphinx
 types-cryptography==3.3.23.2
     # via types-jwt
-types-docutils==0.21.0.20241005
+types-docutils==0.21.0.20241128
     # via -r .typing.in
 types-jwt==0.1.3
     # via -r .typing.in
@@ -32,7 +66,7 @@ typing-extensions==4.12.2
     # via
     #   -r .typing.in
     #   mypy
-urllib3==2.2.3
+urllib3==2.3.0
     # via
     #   requests
     #   responses
